fix: lroom golang version

+++ .dotfiles/zenhan/clean.sh
@@ -1,6 +1,10 @@
 #!/usr/bin/env zsh
 
-BINARY="zenhan.exe"
-WIN_BIN="/mnt/c/Users/$(powershell.exe '$env:USERNAME' | tr -d '\r')/Program Files/zenhan"
+if [[ "$(uname -r)" != *microsoft* ]]; then
+  exit 0
+fi
+
+DIRECTORY="zenhan"
+WIN_BIN="/mnt/c/Users/$(powershell.exe '$env:USERNAME' | tr -d '\r')/Program Files/${DIRECTORY}"
 
 rm -rf ${WIN_BIN}
